# Application/openapi.json
{
  "components": {
    "schemas": {
...
      "CrustIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CrustIn",
        "type": "object"
      },
      "CrustOut": {
        "properties": {
          "crust_id": {
            "title": "Crust Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "crust_id",
          "name"
        ],
        "title": "CrustOut",
        "type": "object"
      },
      "CrustPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CrustPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SizeIn": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_modifier_vnd": {
            "default": 0,
            "title": "Price Modifier Vnd",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "SizeIn",
        "type": "object"
      },
      "SizeOut": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_modifier_vnd": {
            "title": "Price Modifier Vnd",
            "type": "integer"
          },
          "size_id": {
            "title": "Size Id",
            "type": "integer"
          }
        },
        "required": [
          "size_id",
          "name",
          "price_modifier_vnd"
        ],
        "title": "SizeOut",
        "type": "object"
      },
      "SizePatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "price_modifier_vnd": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Modifier Vnd"
          }
        },
        "title": "SizePatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PizzaHUST API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/crusts": {
      "get": {
        "operationId": "list_crusts_api_admin_crusts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CrustOut"
                  },
                  "title": "Response List Crusts Api Admin Crusts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Crusts",
        "tags": [
          "admin-options"
        ]
      },
      "post": {
        "operationId": "create_crust_api_admin_crusts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrustIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrustOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Crust",
        "tags": [
          "admin-options"
        ]
      }
    },
    "/api/admin/crusts/{crust_id}": {
      "delete": {
        "operationId": "delete_crust_api_admin_crusts__crust_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "crust_id",
            "required": true,
            "schema": {
              "title": "Crust Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Crust",
        "tags": [
          "admin-options"
        ]
      },
      "patch": {
        "operationId": "patch_crust_api_admin_crusts__crust_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "crust_id",
            "required": true,
            "schema": {
              "title": "Crust Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrustPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrustOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Crust",
        "tags": [
          "admin-options"
        ]
      }
    },
...
    "/api/admin/sizes": {
      "get": {
        "operationId": "list_sizes_api_admin_sizes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SizeOut"
                  },
                  "title": "Response List Sizes Api Admin Sizes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sizes",
        "tags": [
          "admin-options"
        ]
      },
      "post": {
        "operationId": "create_size_api_admin_sizes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SizeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Size",
        "tags": [
          "admin-options"
        ]
      }
    },
    "/api/admin/sizes/{size_id}": {
      "delete": {
        "operationId": "delete_size_api_admin_sizes__size_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "size_id",
            "required": true,
            "schema": {
              "title": "Size Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Size",
        "tags": [
          "admin-options"
        ]
      },
      "patch": {
        "operationId": "patch_size_api_admin_sizes__size_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "size_id",
            "required": true,
            "schema": {
              "title": "Size Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SizePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SizeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Size",
        "tags": [
          "admin-options"
        ]
      }
    },
...
  }
}